% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{add_dict_matches}
\alias{add_dict_matches}
\title{add_dict_matches}
\usage{
add_dict_matches(
  x,
  col_map = c(text = "text"),
  dicts = c("kantro", "iso3166"),
  matchTypes = c("exact", "contains", "starts"),
  matchCaseVariations = TRUE,
  matchFirstN = Inf,
  addToDF = TRUE
)
}
\arguments{
\item{x}{input data frame}

\item{col_map}{column to match against}

\item{dicts}{list of dictionaries}

\item{matchTypes}{match whole string, start, or anywhere in string}

\item{matchCaseVariations}{match title and all caps versions of dictionary keys}

\item{matchFirstN}{match against partial dictionary keys of at least first N tokens}
}
\value{
tibble with list of measures in a column
}
\description{
add_dict_matches
}
\examples{
[TODO]

}
