% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{add_gender_predictions}
\alias{add_gender_predictions}
\title{add_gender_predictions}
\usage{
add_gender_predictions(
  x,
  col_map = c(given = "given", year = "", country = ""),
  dicts = c("kantro"),
  save_api_results = TRUE,
  fuzzy_match = TRUE,
  year_adjust = TRUE
)
}
\arguments{
\item{col_map}{matching columns to impute}

\item{dicts}{list of dictionaries}

\item{save_api_results}{save api results}

\item{fuzzy_match}{employ fuzzy matching}

\item{year_adjust}{year adjustments}
}
\value{
tibble with imputed gender as probability
}
\description{
add_gender_predictions
}
\examples{
[TODO]
}
