% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{add_local_dict}
\alias{add_local_dict}
\title{Title}
\usage{
add_local_dict(data, name = "", save = FALSE)
}
\arguments{
\item{data}{tibble}

\item{name}{dictionary name}

\item{save}{permanently save}
}
\value{
logical success
}
\description{
Title
}
\examples{
[TODO]
}
