% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{class_estimate}
\alias{class_estimate}
\title{class_estimate}
\usage{
class_estimate(
  x,
  simplify_output = "tidy",
  estimates = c("mean", "sd", "unc"),
  ci_limit = 0.05
)
}
\arguments{
\item{x}{data frame with instrumented by impute}

\item{simplify}{simplify return values}
}
\value{
standard error value of each
}
\description{
class_estimate
}
\examples{
[TODO]
}
