% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{class_mean}
\alias{class_mean}
\title{class_mean}
\usage{
class_mean(x, simplify_output = "scalar")
}
\arguments{
\item{x}{(grouped) data frame with instrumented by impute}

\item{simplify}{return values}
}
\value{
mean value of each
}
\description{
class_mean
}
\examples{
[TODO]
}
