% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{gender_ci}
\alias{gender_ci}
\title{Title}
\usage{
gender_ci(x, cl = 0.05, simplify = FALSE)
}
\arguments{
\item{x}{vector of  pr_f instrumented by impute}

\item{cl}{confidence limit for interval}

\item{simplify}{simplify return values --return pr_f}
}
\value{
mean tibble of upper lower values
}
\description{
Title
}
\examples{
[TODO]
}
