% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{gender_mean}
\alias{gender_mean}
\title{gender_mean}
\usage{
gender_mean(x, simplify_output = "scalar")
}
\arguments{
\item{x}{(grouped) data frame with instrumented by impute}

\item{simplify}{return values}
}
\value{
mean value of each
}
\description{
gender_mean
}
\examples{
[TODO]
}
