% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{gender_sample}
\alias{gender_sample}
\title{Title}
\usage{
gender_sample(x, simplify = FALSE, n = 1)
}
\arguments{
\item{x}{data frame with instrumented by impute}

\item{n}{number of imputations to draw}
}
\value{
list of imputation of gender
}
\description{
Title
}
\examples{
[TODO]
}
