% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{gender_se}
\alias{gender_se}
\title{Title}
\usage{
gender_se(x, simplify = TRUE)
}
\arguments{
\item{x}{data frame with instrumented by impute}

\item{simplify}{return values}
}
\value{
standard error value of each
}
\description{
Title
}
\examples{
[TODO]
}
