% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{impute_gender}
\alias{impute_gender}
\title{Title}
\usage{
impute_gender(
  x,
  col_map = c(given = "given", year = "", country = ""),
  dicts = c("kantro"),
  dict_api = "",
  save_api = TRUE,
  fuzzy_match = TRUE,
  year_adjust = TRUE
)
}
\arguments{
\item{year_adjust}{}
}
\description{
Title
}
