% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso3166}
\alias{iso3166}
\title{ISO 3166 country codes}
\format{
\subsection{\code{iso3166}}{

A data frame 252 observation and 3 columns
\describe{
\item{country_name}{Country name}
\item{country_2d}{2 letter ISO country codes}
\item{country_num}{country numeric code}
...
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/ISO_3166-2}
}
\usage{
iso3166
}
\description{
Country codes and names
}
\keyword{datasets}
