% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kantro}
\alias{kantro}
\title{Kantrowitz Corpus of Male and Female names from NLTK system}
\format{
\subsection{\code{kantro}}{

A data frame XXX observation and 2 columns
\describe{
\item{pr_f}{proportion female}
\item{given}{given name}
...
}
}
}
\source{
\url{https://www.nltk.org/book/ch06.html}
}
\usage{
kantro
}
\description{
Country codes and names
}
\keyword{datasets}
