% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{load_dict}
\alias{load_dict}
\title{Title}
\usage{
load_dict(name, force = FALSE)
}
\arguments{
\item{name}{name parameter}

\item{force}{Force reload}
}
\value{
logical success
}
\description{
Title
}
\examples{
[TODO]
}
