% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{manage_local_dict}
\alias{manage_local_dict}
\title{manage_local_dictionary}
\usage{
manage_local_dict(
  x,
  name = "local_1",
  description = "a local dictionary",
  delete = FALSE,
  force = FALSE
)
}
\arguments{
\item{name}{dictionary name}

\item{data}{tibble}

\item{save}{permanently save}
}
\value{
logical success
}
\description{
manage_local_dictionary
}
\examples{
[TODO]
}
