% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{show_dict}
\alias{show_dict}
\title{Title}
\usage{
show_dict(name)
}
\arguments{
\item{name}{name of dictionary}
}
\value{
tibble containing selected directionary
}
\description{
Title
}
\examples{
[TODO]
}
