% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengender.R
\name{manage_local_dicts}
\alias{manage_local_dicts}
\title{manage_local_dicts}
\usage{
manage_local_dicts(
  x,
  name = "local_1",
  description = "a local dictionary",
  delete = FALSE,
  force = FALSE,
  min_obs = options("opengender.dict.minsize")[[1]],
  domain = "gender",
  version = "1",
  levels = NULL
)
}
\arguments{
\item{name}{dictionary name}

\item{description}{description}

\item{delete}{permanently delete}

\item{force}{force replacing existing added dictionary}

\item{min_obs}{mininum observations per cell to use when normalizing}

\item{data}{tibble}
}
\value{
logical success
}
\description{
manage_local_dicts
}
\examples{
[TODO]
}
